pyproject.toml
[project]
name = "friends-scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py-self>=2.0.1",
    "mypy>=1.15.0",
    "python-dotenv>=1.1.0",
]
